Structured the code base
Created a public resources class, which stores all the Textures and Sprites for easy access throughout the code. Made a DefaultScreen class in screens package, and all the following up screens will inherit it. Renamed MainMenu.java to MainScreen.java .

There is an error while running the code due to the show() method in MainScreen.java